+++ src/components/index.ts
@@ -1,0 +1,7 @@
+export { default as CommentListItem } from './CommentListItem';
+export { default as EmptyList } from './EmptyList';
+export { default as Error } from './Error';
+export { default as Header } from './Header';
+export { default as Loader } from './Loader';
+export { default as PostList } from './PostList';
+export { default as PostListItem } from './PostListItem';
